Add files via upload
My solution for using the diff function when in a bin is to create a global boolean variable that will be set to true at the end of the diff function. When the diff function is called, there is a conditional statement that checks for the boolean variable and it will perform differently if that is the case.